Change preferences imports after moving.
