+++ spyder/preferences/runconfig.py
@@ -21,7 +21,7 @@ from qtpy.QtWidgets import (QButtonGroup, QCheckBox, QComboBox, QDialog,
 # Local imports
 from spyder.config.base import _
 from spyder.config.main import CONF
-from spyder.plugins.configdialog import GeneralConfigPage
+from spyder.preferences.configdialog import GeneralConfigPage
 from spyder.py3compat import getcwd, to_text_string
 from spyder.utils import icon_manager as ima
