+++ spyder/preferences/shortcuts.py
@@ -27,7 +27,7 @@ from qtpy.QtWidgets import (QAbstractItemView, QApplication, QDialog,
 from spyder.config.base import _, debug_print
 from spyder.config.gui import (get_shortcut, iter_shortcuts,
                                reset_shortcuts, set_shortcut)
-from spyder.plugins.configdialog import GeneralConfigPage
+from spyder.preferences.configdialog import GeneralConfigPage
 from spyder.utils import icon_manager as ima
 from spyder.utils.qthelpers import get_std_icon
 from spyder.utils.stringmatching import get_search_scores, get_search_regex
