+++ spyder/widgets/projects/configdialog.py
@@ -10,7 +10,7 @@
 from qtpy.QtWidgets import QGroupBox, QVBoxLayout
 
 from spyder.config.base import _
-from spyder.plugins.configdialog import ConfigDialog, GeneralConfigPage
+from spyder.preferences.configdialog import ConfigDialog, GeneralConfigPage
 from spyder.utils.qthelpers import get_icon
 from spyder.config.user import NoDefault
 from spyder.widgets.projects import EmptyProject
