+++ spyder_profiler/profiler.py
@@ -19,7 +19,7 @@ from qtpy.QtWidgets import QGroupBox, QLabel, QVBoxLayout
 from spyder.config.base import get_translation
 from spyder.api.plugins import SpyderPluginWidget
 from spyder.api.preferences import PluginConfigPage
-from spyder.plugins.runconfig import get_run_configuration
+from spyder.preferences.runconfig import get_run_configuration
 from spyder.utils import icon_manager as ima
 from spyder.utils.qthelpers import create_action
 from .widgets.profilergui import (ProfilerWidget, is_profiler_installed)
